requirements.txt
flask==3.0.3
gunicorn==22.0.0
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
ultralytics==8.3.179
matplotlib==3.8.4
seaborn==0.13.2
